// src / graphql / generated.ts
export type Maybe<T> = T | null;
// ...
export type Scalars = {
  ID: { input: string; output: string; }
  String: { input: string; output: string; }
  Boolean: { input: boolean; output: boolean; }
  Int: { input: number; output: number; }
  Float: { input: number; output: number; }
  BlockchainData: { input: any; output: any; }
  BroadcastId: { input: any; output: any; }
  ChainId: { input: any; output: any; }
  CollectModuleData: { input: any; output: any; }
  ContentEncryptionKey: { input: any; output: any; }
  ContractAddress: { input: any; output: any; }
  CreateHandle: { input: any; output: any; }
  Cursor: { input: any; output: any; }
  DataAvailabilityId: { input: any; output: any; }
  DateTime: { input: any; output: any; }
  EncryptedValueScalar: { input: any; output: any; }
  Ens: { input: any; output: any; }
  EthereumAddress: { input: any; output: any; }
  FollowModuleData: { input: any; output: any; }
  Handle: { input: any; output: any; }
  HandleClaimIdScalar: { input: any; output: any; }
  ImageSizeTransform: { input: any; output: any; }
  InternalPublicationId: { input: any; output: any; }
  IpfsCid: { input: any; output: any; }
  Jwt: { input: any; output: any; }
  LimitScalar: { input: any; output: any; }
  Locale: { input: any; output: any; }
  Markdown: { input: any; output: any; }
  MimeType: { input: any; output: any; }
  NftGalleryId: { input: any; output: any; }
  NftGalleryName: { input: any; output: any; }
  NftOwnershipId: { input: any; output: any; }
  Nonce: { input: any; output: any; }
  NotificationId: { input: any; output: any; }
  ProfileId: { input: any; output: any; }
  ProfileInterest: { input: any; output: any; }
  ProxyActionId: { input: any; output: any; }
  PublicationId: { input: any; output: any; }
  PublicationTag: { input: any; output: any; }
  PublicationUrl: { input: any; output: any; }
  ReactionId: { input: any; output: any; }
  ReferenceModuleData: { input: any; output: any; }
  Search: { input: any; output: any; }
  Signature: { input: any; output: any; }
  Sources: { input: any; output: any; }
  TimestampScalar: { input: any; output: any; }
  TokenId: { input: any; output: any; }
  TxHash: { input: any; output: any; }
  TxId: { input: any; output: any; }
  UnixTimestamp: { input: any; output: any; }
  Url: { input: any; output: any; }
  Void: { input: any; output: any; }
};
// ...
export type CanCommentResponse = {
  __typename?: 'CanCommentResponse';
  result: Scalars['Boolean']['output'];
};

export type CanDecryptResponse = {
  __typename?: 'CanDecryptResponse';
  extraDetails?: Maybe<Scalars['String']['output']>;
  reasons?: Maybe<Array<DecryptFailReason>>;
  result: Scalars['Boolean']['output'];
};

export type CanMirrorResponse = {
  __typename?: 'CanMirrorResponse';
  result: Scalars['Boolean']['output'];
};
// ...
export type CollectModule = AaveFeeCollectModuleSettings | Erc4626FeeCollectModuleSettings | FeeCollectModuleSettings | FreeCollectModuleSettings | LimitedFeeCollectModuleSettings | LimitedTimedFeeCollectModuleSettings | MultirecipientFeeCollectModuleSettings | RevertCollectModuleSettings | SimpleCollectModuleSettings | TimedFeeCollectModuleSettings | UnknownCollectModuleSettings;
// ...
export type Comment = {
  __typename?: 'Comment';
  /** ID of the source */
  appId?: Maybe<Scalars['Sources']['output']>;
  bookmarked: Scalars['Boolean']['output'];
  canComment: CanCommentResponse;
  canDecrypt: CanDecryptResponse;
  canMirror: CanMirrorResponse;
  /** The collect module */
  collectModule: CollectModule;
  /** The contract address for the collect nft.. if its null it means nobody collected yet as it lazy deployed */
  collectNftAddress?: Maybe<Scalars['ContractAddress']['output']>;
  /** Who collected it, this is used for timeline results and like this for better caching for the client */
  collectedBy?: Maybe<Wallet>;
  /** Which comment this points to if its null the pointer too deep so do another query to find it out */
  commentOn?: Maybe<Publication>;
  /** The date the post was created on */
  createdAt: Scalars['DateTime']['output'];
  /** The data availability proofs you can fetch from */
  dataAvailabilityProofs?: Maybe<Scalars['String']['output']>;
  /** This will bring back the first comment of a comment and only be defined if using `publication` query and `commentOf` */
  firstComment?: Maybe<Comment>;
  hasCollectedByMe: Scalars['Boolean']['output'];
  /** If the publication has been hidden if it has then the content and media is not available */
  hidden: Scalars['Boolean']['output'];
  /** The internal publication id */
  id: Scalars['InternalPublicationId']['output'];
  /** Indicates if the publication is data availability post */
  isDataAvailability: Scalars['Boolean']['output'];
  /** Indicates if the publication is gated behind some access criteria */
  isGated: Scalars['Boolean']['output'];
  /** The top level post/mirror this comment lives on */
  mainPost: MainPostReference;
  /** The metadata for the post */
  metadata: MetadataOutput;
  mirrors: Array<Scalars['InternalPublicationId']['output']>;
  notInterested: Scalars['Boolean']['output'];
  /** The on chain content uri could be `ipfs://` or `https` */
  onChainContentURI: Scalars['String']['output'];
  /** The profile ref */
  profile: Profile;
  /** Comment ranking score */
  rankingScore?: Maybe<Scalars['Float']['output']>;
  reaction?: Maybe<ReactionTypes>;
  /** The reference module */
  referenceModule?: Maybe<ReferenceModule>;
  /** The publication stats */
  stats: PublicationStats;
};
// ...
export type MainPostReference = Mirror | Post;
// ...
export type MetadataOutput = {
  __typename?: 'MetadataOutput';
  /** The main focus of the publication */
  animatedUrl?: Maybe<Scalars['Url']['output']>;
  /** The attributes */
  attributes: Array<MetadataAttributeOutput>;
  /** This is the metadata content for the publication, should be markdown */
  content?: Maybe<Scalars['Markdown']['output']>;
  /** The content warning for the publication */
  contentWarning?: Maybe<PublicationContentWarning>;
  /** The image cover for video/music publications */
  cover?: Maybe<MediaSet>;
  /** This is the metadata description */
  description?: Maybe<Scalars['Markdown']['output']>;
  /** The publication's encryption params in case it's encrypted */
  encryptionParams?: Maybe<EncryptionParamsOutput>;
  /** This is the image attached to the metadata and the property used to show the NFT! */
  image?: Maybe<Scalars['Url']['output']>;
  /** The locale of the publication,  */
  locale?: Maybe<Scalars['Locale']['output']>;
  /** The main focus of the publication */
  mainContentFocus: PublicationMainFocus;
  /** The images/audios/videos for the publication */
  media: Array<MediaSet>;
  /** The metadata name */
  name?: Maybe<Scalars['String']['output']>;
  /** The tags for the publication */
  tags: Array<Scalars['String']['output']>;
};

/** The social mirror */
export type Mirror = {
  __typename?: 'Mirror';
  /** ID of the source */
  appId?: Maybe<Scalars['Sources']['output']>;
  bookmarked: Scalars['Boolean']['output'];
  canComment: CanCommentResponse;
  canDecrypt: CanDecryptResponse;
  canMirror: CanMirrorResponse;
  /** The collect module */
  collectModule: CollectModule;
  /** The contract address for the collect nft.. if its null it means nobody collected yet as it lazy deployed */
  collectNftAddress?: Maybe<Scalars['ContractAddress']['output']>;
  /** The date the post was created on */
  createdAt: Scalars['DateTime']['output'];
  /** The data availability proofs you can fetch from */
  dataAvailabilityProofs?: Maybe<Scalars['String']['output']>;
  hasCollectedByMe: Scalars['Boolean']['output'];
  /** If the publication has been hidden if it has then the content and media is not available */
  hidden: Scalars['Boolean']['output'];
  /** The internal publication id */
  id: Scalars['InternalPublicationId']['output'];
  /** Indicates if the publication is data availability post */
  isDataAvailability: Scalars['Boolean']['output'];
  /** Indicates if the publication is gated behind some access criteria */
  isGated: Scalars['Boolean']['output'];
  /** The metadata for the post */
  metadata: MetadataOutput;
  /** The mirror publication */
  mirrorOf: MirrorablePublication;
  notInterested: Scalars['Boolean']['output'];
  /** The on chain content uri could be `ipfs://` or `https` */
  onChainContentURI: Scalars['String']['output'];
  /** The profile ref */
  profile: Profile;
  reaction?: Maybe<ReactionTypes>;
  /** The reference module */
  referenceModule?: Maybe<ReferenceModule>;
  /** The publication stats */
  stats: PublicationStats;
};
// ...
export type MirrorablePublication = Comment | Post;
// ...
export type Post = {
  __typename?: 'Post';
  /** ID of the source */
  appId?: Maybe<Scalars['Sources']['output']>;
  bookmarked: Scalars['Boolean']['output'];
  canComment: CanCommentResponse;
  canDecrypt: CanDecryptResponse;
  canMirror: CanMirrorResponse;
  /** The collect module */
  collectModule: CollectModule;
  /** The contract address for the collect nft.. if its null it means nobody collected yet as it lazy deployed */
  collectNftAddress?: Maybe<Scalars['ContractAddress']['output']>;
  /**
   * Who collected it, this is used for timeline results and like this for better caching for the client
   * @deprecated use `feed` query, timeline query will be killed on the 15th November. This includes this field.
   */
  collectedBy?: Maybe<Wallet>;
  /** The date the post was created on */
  createdAt: Scalars['DateTime']['output'];
  /** The data availability proofs you can fetch from */
  dataAvailabilityProofs?: Maybe<Scalars['String']['output']>;
  hasCollectedByMe: Scalars['Boolean']['output'];
  /** If the publication has been hidden if it has then the content and media is not available */
  hidden: Scalars['Boolean']['output'];
  /** The internal publication id */
  id: Scalars['InternalPublicationId']['output'];
  /** Indicates if the publication is data availability post */
  isDataAvailability: Scalars['Boolean']['output'];
  /** Indicates if the publication is gated behind some access criteria */
  isGated: Scalars['Boolean']['output'];
  /** The metadata for the post */
  metadata: MetadataOutput;
  mirrors: Array<Scalars['InternalPublicationId']['output']>;
  notInterested: Scalars['Boolean']['output'];
  /** The on chain content uri could be `ipfs://` or `https` */
  onChainContentURI: Scalars['String']['output'];
  /** The profile ref */
  profile: Profile;
  reaction?: Maybe<ReactionTypes>;
  /** The reference module */
  referenceModule?: Maybe<ReferenceModule>;
  /** The publication stats */
  stats: PublicationStats;
};
// ...
export type Publication = Comment | Mirror | Post;
// ...
export type PublicationStats = {
  __typename?: 'PublicationStats';
  commentsTotal: Scalars['Int']['output'];
  /** The publication id */
  id: Scalars['InternalPublicationId']['output'];
  /** The total amount of collects */
  totalAmountOfCollects: Scalars['Int']['output'];
  /** The total amount of comments */
  totalAmountOfComments: Scalars['Int']['output'];
  /** The total amount of mirrors */
  totalAmountOfMirrors: Scalars['Int']['output'];
  /** The total amount of bookmarks */
  totalBookmarks: Scalars['Int']['output'];
  /** The total amount of upvotes */
  totalDownvotes: Scalars['Int']['output'];
  /** The total amount of downvotes */
  totalUpvotes: Scalars['Int']['output'];
};
// ...
export enum ReactionTypes {
  Downvote = 'DOWNVOTE',
  Upvote = 'UPVOTE'
}
// ...
export type ReferenceModule = DegreesOfSeparationReferenceModuleSettings | FollowOnlyReferenceModuleSettings | UnknownReferenceModuleSettings;
// ...
export type Wallet = {
  __typename?: 'Wallet';
  address: Scalars['EthereumAddress']['output'];
  /** The default profile for the wallet for now it is just their first profile, this will be the default profile they picked soon enough */
  defaultProfile?: Maybe<Profile>;
};